new BlocProvider class
