+++ lib/models/_models.dart
@@ -1,4 +1,3 @@
-export 'abstracts.dart';
 export 'api_id.dart';
 export 'theater.dart';
 export 'movie.dart';
