+++ lib/utils/_utils.dart
@@ -1,3 +1,5 @@
 export 'extensions.dart';
 export 'utils.dart';
 export 'message.dart';
+export 'abstracts/bloc_provider.dart';
+export 'abstracts/disposable.dart';
