+++ lib/utils/abstracts/disposable.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/foundation.dart';
+
+mixin Disposable {
+  bool isDisposed = false;
+
+  @mustCallSuper
+  void dispose() {
+    isDisposed = true;
+  }
+}
